Separate `WakeUpDetectorProtocol`
